myapp/requirements.txt
dotenv
aioredis==1.2.0
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
certifi==2019.9.11
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.5
django-jenkins==0.110.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
javascript==2015.32.59
msgpack==0.6.1
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.3
zope.interface==4.6.0
